Cambio ints a float en EJercicio4

+++ Ejercicio4.java
@@ -2,7 +2,7 @@ public class Ejercicio4{
 	
 	public static void main(String[] args) {
 		
-		int n,m;
+		float n,m;
 		n=5;
 		m=3;
 		System.out.println("La division es " + n/m);
